Move utils/strings.go into it's own package
